bit depth changed on one it demo to run, setup.py cleaned from tensorflow.txt as it is removed from repo with previous commits

+++ aydin/it/demo/n2s_nn_3D_small_org.py
@@ -9,8 +9,6 @@ from aydin.features.tiled.tiled_features import TiledFeatureGenerator
 from aydin.io import io
 from aydin.io.datasets import examples_single
 from aydin.it.it_classic import ImageTranslatorClassic
-from aydin.regression.clnn import CLNNRegressor
-from aydin.regression.gbm import GBMRegressor
 from aydin.regression.nn import NNRegressor
 
 
@@ -28,7 +26,7 @@ def demo():
     image = rescale_intensity(image, in_range='image', out_range=(0, 1))
 
     generator = TiledFeatureGenerator(
-        FastMultiscaleConvolutionalFeatures(max_level=6, dtype=numpy.uint8)
+        FastMultiscaleConvolutionalFeatures(max_level=6, dtype=numpy.float32)
     )
     regressor = NNRegressor(depth=6, max_epochs=40, patience=10)
     it = ImageTranslatorClassic(
